cleaning up directory
