--- bin/config_sample
@@ -1,10 +1,0 @@
-127.0.0.1
-#minechannel
-txtminebot
-admin_nick
-10
-lang/commands.txt
-lang/interp-pos.txt
-lang/interp-neu.txt
-lang/interp-neg.txt
-lang/wham.txt
