+++ bin/ircconfig_sample
@@ -1,0 +1,4 @@
+127.0.0.1
+#minechannel
+bot_nick
+admin_nick
