--- bin/standardrates
@@ -1,9 +1,0 @@
-40
-20
-15
-10
-6
-4
-4
-1
-100
