+++ bin/util.py
@@ -51,3 +51,20 @@ def pretty_time(time):
             return p.no("minute", m) + " and " + p.no("second", s)
     else:
         return p.no("second", s)
+
+def format_message(message):
+    # this came from tilde.town
+
+    #pattern = r'^:.*\!~(.*)@.* (.*) (.*) :(.*)'
+    pattern = r'^:.*\!(.*)@.* (.*) (.*) :(.*)'
+    now = int(time.time())
+    matches = re.match(pattern, message)
+    if not matches:
+        return ''
+
+    nick    = matches.group(1).strip()
+    command = matches.group(2).strip()
+    channel = matches.group(3).strip()
+    message = matches.group(4).strip()
+
+    return "%s\t%s\t%s\t%s\t%s" % (now, nick, command, channel, message)
